Changed recommended install order
I was encountering errors in windows that had to do with the order that I executed the commands as per the readme file. Git automatically adds a different end of line character to all text files when downloaded in Windows. The current readme instructions suggest that we do a git fetch command in windows before installing the vagrant box. We then read those windows files from the shared directory on the vagrant box. This caused errors when trying to execute commands in the vagrant ssh dialog. 

These changes ask the user to clone the repos inside the vagrant box to avoid end of line errors brought on by git Windows. We still clone the items to the shared directory so that files can be easily added from the host OS. 

I also added a note to the "Building the Virtualenv" section to run this as administrator in Windows because of symlink permission errors caused during the NPM install `sudo npm install npm`